roomservice.py: fetch our devices
This was a bit hard.
The roomservice now looks for the device sources in our repo, whuch means that only those can be builded now,
maybe we can find a way that it looks for devices on CM repo too if we dont have it in our repo.


##########################################################
VERY IMPORTANT!!:

Every repo in the device dependencies must now contain:

remote=""    AND
revision=""

in order to find the repo we are looking for.

If we track repoÂ´s from out Github and the default lollipop-5.0 branch we dont need to add these to things to the repo 
in the dependencies
##########################################################

Change-Id: I4c412ae50563695b827cd4848352972cf74edf4d
Signed-off-by: noname55804 <marlin2201@gmail.com>